# 03-Azure/01-04-AI/01_Agentic_SDLC/src/api-ts/api-swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Express API with Swagger",
        "version": "1.0.0",
        "description": "REST API documentation using Swagger/OpenAPI"
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/api/headquarters": {
            "get": {
                "summary": "Returns all headquarters",
                "tags": [
                    "Headquarters"
                ],
                "responses": {
                    "200": {
                        "description": "List of all headquarters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Headquarters"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new headquarters",
                "tags": [
                    "Headquarters"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Headquarters"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Headquarters created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Headquarters"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/headquarters/{id}": {
            "get": {
                "summary": "Get a headquarters by ID",
                "tags": [
                    "Headquarters"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Headquarters ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Headquarters found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Headquarters"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Headquarters not found"
                    }
                }
            },
            "put": {
                "summary": "Update a headquarters",
                "tags": [
                    "Headquarters"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Headquarters ID"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Headquarters"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Headquarters updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Headquarters"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Headquarters not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete a headquarters",
                "tags": [
                    "Headquarters"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Headquarters ID"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Headquarters deleted successfully"
                    },
                    "404": {
                        "description": "Headquarters not found"
                    }
                }
            }
        },
...
        "/api/headquarters/{id}/label": {
            "get": {
                "summary": "Get headquarters label by ID",
                "tags": [
                    "Headquarters"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Headquarters ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Headquarters label created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "label": {
                                            "type": "string",
                                            "description": "Formatted location label"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Headquarters not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Headquarters": {
                "type": "object",
                "required": [
                    "headquartersId",
                    "name"
                ],
                "properties": {
                    "headquartersId": {
                        "type": "integer",
                        "description": "The unique identifier for the headquarters"
                    },
                    "name": {
                        "type": "string",
                        "description": "The name of the headquarters"
                    },
                    "address": {
                        "type": "string",
                        "description": "Main office address of the headquarters"
                    },
                    "phone": {
                        "type": "string",
                        "description": "Contact phone number for the headquarters"
                    },
                    "description": {
                        "type": "string",
                        "description": "Additional details about the headquarters"
                    },
                    "contactPerson": {
                        "type": "string",
                        "description": "Name of the primary contact person"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "description": "Contact email for the headquarters"
                    },
                    "city": {
                        "type": "string",
                        "description": "City where the headquarters is located"
                    },
                    "country": {
                        "type": "string",
                        "description": "Country where the headquarters is located"
                    },
                    "floorCount": {
                        "type": "integer",
                        "description": "Number of floors in the headquarters building"
                    },
                    "capacity": {
                        "type": "integer",
                        "description": "Total capacity of the headquarters"
                    }
                }
            },
...
        }
    },
...
}